dj_pro/requirement.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.0
Django==3.2.4
django-cors-headers==3.10.1
django-filter==21.1
django-simpleui==2022.1
django-treebeard==4.3
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
idna==3.3
Pillow==8.4.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.12.0
pycryptodomex==3.12.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-alipay-sdk==1.8.0
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tencentcloud-sdk-python==3.0.547
tzdata==2021.5
urllib3==1.26.7
